rename utils to db_models
